+++ packages/eyes-selenium/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- support edgelegacy, edgechromium, and edgechromium-one-version-back
 
 ## 4.34.0 - 2020/4/26